requirements.txt
apache-airflow==1.10.9
pandas==0.25.3
falcon==2.0.0
pymongo==3.10.1
pytest==5.4.1
pmdarima==1.5.3
fbprophet==0.6
gunicorn==19.7.1